APN: Remove identical MVNO entries
Starting M-Android, APN's that are not unique are
clubbed in one entry in the telephony db. "name"
field is not part of the unique set and you end
up with a weird MVNO name as the APN name for
T-Mobile. Which ever is the last entry will be the
final name. In case of TMO it became "FMP Communications"

Change-Id: Ia86d5c72f36d32315ef1e2469fdd364869160730
Signed-off-by: bgos10 <bishalgoswami@gmail.com>